+++ greattunes/__init__.py
@@ -5,11 +5,11 @@ from ._acq_func import AcqFunction
 from ._version import __version__
 
 
-class CreativeProject(Initializers, AcqFunction):
+class TuneSession(Initializers, AcqFunction):
     """
     user-facing functionality to balance exploration and exploitation for creative projects.
     Note: Initializers is a child of Validators class (from ._validators.py) so all validator methods are available to
-    CreativeProject
+    TuneSession
     """
 
     # Initialize class instance
@@ -17,7 +17,7 @@ class CreativeProject(Initializers, AcqFunction):
         self,
         covars,
         model="SingleTaskGP",
-        acq_func="EI",
+        acq_func="ExpectedImprovement",
         random_start=True,
         random_step_cadence=10,
         **kwargs,
@@ -31,23 +31,29 @@ class CreativeProject(Initializers, AcqFunction):
         variable. Currently only allows for continuous variables. Aspiration: Data type of input will be preserved and
         code be adapted to accommodate both integer and categorical variables.
         :param model str: sets surrogate model (currently allow "SingleTaskGP" and "Custom")
-        :param acq_func (str): sets acquisition function (currently allows only "EI")
+        :param acq_func (str): sets acquisition function (list of available acquisition functions available as
+        self.ACQ_FUNC_LIST)
         :param random_start (bool): determines whether to start from random. If set to True and previously obtained
         data provided via train_X and train_Y, there will still be random samples at start
         :param random_step_cadence (int/None): proposes a random datapoint (not determined by Bayesian optimization)
         every 'random_step_cadence' iterations. Runs no random datapoints if set to None. This helps in finding the
         global optimum.
         :kwargs
-            - train_X (torch.tensor of dtype=torch.double): design matrix of covariates (batch_shape X num_obs X
-            num_training_features OR num_obs X num_training_features)
-            - train_Y (torch.tensor of dtype=torch.double): observations (batch_shape X num_obs X num_output_models
-            [allows for batched models] OR num_obs X num_output_models)
+            - beta (float): tradeoff parameter for acquisition functions UpperConfidenceBound, qUpperConfidenceBound
             - kernel (str): kernel used to define custom GP models (not currently in use)
             - nu (float): kernel parameter for Matern kernel
+            - num_fantasies (int): number of realizations for generating estimates for acquisition functions
+            qKnowledgeGradient, NoisyExpectedImprovement
             - num_initial_random (int): number of initial random points. Only changes anything if 'random_start' is
             True
             - random_sampling_method (str): sampling method for random points. Options: "random" and "latin_hcs" (latin
             hypercube sampling)
+            - sampler (BoTorch acqusition sampler object): sampler for picking datapoints for series of acquisition
+            functions. Details on available samplers on https://botorch.org/api/sampling.html.
+            - train_X (torch.tensor of dtype=torch.double): design matrix of covariates (batch_shape X num_obs X
+            num_training_features OR num_obs X num_training_features)
+            - train_Y (torch.tensor of dtype=torch.double): observations (batch_shape X num_obs X num_output_models
+            [allows for batched models] OR num_obs X num_output_models)
         """
 
         # === Computational settings ===
@@ -57,7 +63,47 @@ class CreativeProject(Initializers, AcqFunction):
         # data type for tensors
         self.dtype = torch.double
 
-        # === Data initialization ===
+        # === Parameter initialization ===
+
+        # define kernel --- THIS SHOULD BE DONE MORE ELEGANTLY. If not present as input, these will be set to None
+        self.nu = kwargs.get("nu")  # get value of nu for Matern kernel
+
+        # define acquisition function parameters
+        # number of realizations to average in qKnowledgeGradient, NoisyExpectedImprovement
+        # if no input provided and models requiring it are picked, select a number for it
+        self.num_fantasies = kwargs.get("num_fantasies")
+        if (
+            acq_func in ["qKnowledgeGradient", "NoisyExpectedImprovement"]
+            and self.num_fantasies is None
+        ):
+            self.num_fantasies = 20
+            if acq_func == "qKnowledgeGradient":
+                self.num_fantasies = 64
+
+        # tradeoff parameter for UpperConfidenceBound, qUpperConfidenceBound
+        self.beta = kwargs.get("beta")
+        if (
+            acq_func in ["UpperConfidenceBound", "qUpperConfidenceBound"]
+            and self.beta is None
+        ):
+            self.beta = 0.2
+
+        # sampler for picking data points
+        # if none provided but one is needed, pick a SobolQMCNormal sampler from BoTorch
+        self.sampler = kwargs.get("sampler")
+        list_acqf_sampler = [
+            "qExpectedImprovement",
+            "qNoisyExpectedImprovement",
+            "qProbabilityOfImprovement",
+            "qSimpleRegret",
+            "qUpperConfidenceBound",
+        ]
+        if acq_func in list_acqf_sampler and self.sampler is None:
+            from botorch.sampling import SobolQMCNormalSampler
+
+            self.sampler = SobolQMCNormalSampler(1024)
+
+        # === Model and optimization initialization ===
         # initialize the data (initial guess) and bounds on covars
         (
             self.initial_guess,
@@ -76,14 +122,16 @@ class CreativeProject(Initializers, AcqFunction):
             "response_sampled_iter": 0,
         }
 
-        # define kernel --- THIS SHOULD BE DONE MORE ELEGANTLY. If not present as input, these will be set to None
-        self.nu = kwargs.get("nu")  # get value of nu for Matern kernel
-
         # define acquisition function
         self.acq_func = {
             "type": acq_func,
             "object": None,
         }
+
+        # list available acquisition functions
+        AcqFunction.__init__(
+            self
+        )  # creates ACQ_FUNC_LIST attribute by running constructor of AcqFunc parent class
 
         # define sampling functions. initialize as iterative, which means using ask-tell (either manual or automatic).
         # Will be updated if method "auto" is used (use when sampling function known)
@@ -147,7 +195,7 @@ class CreativeProject(Initializers, AcqFunction):
         if self.nu is not None:
             deep_str += f", nu={self.nu!r}"
 
-        return f"CreativeProject(" + deep_str + f")"
+        return f"TuneSession(" + deep_str + f")"
 
     def __str__(self):
         """
@@ -165,7 +213,7 @@ class CreativeProject(Initializers, AcqFunction):
         if self.nu is not None:
             deep_str += f", nu={self.nu}"
 
-        return f"CreativeProject(" + deep_str + f")"
+        return f"TuneSession(" + deep_str + f")"
 
     # import methods
     from ._campaign import auto, ask, tell